+++ Assets/Scripts/Slime.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Slime : Enemigo
 {
 
-    [SerializeField] private float danhoAtaque;
+    [SerializeField] private int danhoAtaque;
     [SerializeField] private Transform puntoAtaque;
     [SerializeField] private float radioAtaque;
     [SerializeField] private LayerMask queEsDanhable;
